Tags and Products set, warehouse edit glitching

+++ public/js/product.js
@@ -1,12 +1,15 @@
 // add product function, on click inside add product modal.
 const addProduct = async (event) => {
   event.preventDefault();
+  const product_name = document
+    .querySelector('#new-product-name')
+    .value.replace(/\s+/g, '-');
 
   const body = {
-    product_name: $('#new-product-name').val(),
+    product_name: product_name,
     price: $('#new-price').val(),
     stock: parseInt($('#new-stock').val()),
-    category_id: parseInt($('#new-category-id').val()),
+    warehouse_id: parseInt($('#new-warehouse-id').val()),
     tagIds: $('#new-tag-id').val(),
   };
 
@@ -37,25 +40,35 @@ const addProduct = async (event) => {
 };
 
 // edit product function, on click inside edit product modal.
-const editProduct = async (id) => {
+const editProduct = async (event, id) => {
+  event.preventDefault();
+  const product_name = document
+    .querySelector('#product-name')
+    .value.replace(/\s+/g, '-');
 
   const body = {
-    product_name: $('#product-name').val(),
+    product_name: product_name,
     price: $('#price').val(),
     stock: parseInt($('#stock').val()),
-    category_id: parseInt($('#category_id').val()),
-    tagIds: $('#tag_id').val(),
+    warehouse_id: parseInt($('#warehouse-id').val()),
+    tagIds: $('#tag-id').val(),
   };
-
-  console.log(body.tagIds);
-  console.log(body);
-
+  // validate entry of required fields. Could do this inline html, keeping it on client side js for back-end mvp project..
   if (body.product_name === '') {
     alert('You must enter at least one character for a product name');
     return;
   }
 
-  const response = await fetch(`api/products/${id}`, {
+  if (body.price === '') {
+    alert('A product must have a price.');
+    return;
+  }
+  if (body.warehouse_id === '') {
+    alert('You must assign this product to a warehouse.');
+    return;
+  }
+
+  const response = await fetch(`/api/products/${id}`, {
     method: 'PUT',
     headers: {
       'Content-Type': 'application/json',
@@ -65,8 +78,8 @@ const editProduct = async (id) => {
 
   if (response.ok) {
     const myProduct = await response.json();
-    alert(`Edited product' + ${myProduct.product_name}`);
-    document.location.href = '/';
+    alert(`Edited product!`);
+    window.location.href = '/';
   } else {
     const errorObj = await response.json();
     console.log(errorObj.message);
@@ -74,22 +87,26 @@ const editProduct = async (id) => {
 };
 
 // client side js to call our routes from the frontend.
-const deleteProduct = async (id) => {
-  alert('you reached delete product');
-  try {
-    alert(id);
-    const response = await fetch(`/api/products/${id}}`, {
-      method: 'DELETE',
-    });
+const deleteProduct = async (event, id) => {
+  event.preventDefault();
+  if (
+    confirm('This step can not be undone, are you sure you want to delete?')
+  ) {
+    try {
+      alert(id);
+      const response = await fetch(`/api/products/${id}}`, {
+        method: 'DELETE',
+      });
 
-    if (response.ok) {
-      alert(`Product deleted!`);
-      window.location.href = '/';
-    } else {
-      return console.log(response);
+      if (response.ok) {
+        alert(`Product deleted!`);
+        window.location.href = '/';
+      } else {
+        return console.log(response);
+      }
+    } catch (err) {
+      console.log('Unable to delete product.');
     }
-  } catch (err) {
-    console.log('Unable to delete product.');
   }
+  return;
 };
-
